Create audiobook.py
In this example, we first import the gTTS library and the os module. Then, we define the text we want to convert to speech and the language we want to use (in this case, English). We then create an audio object using the gTTS function, passing in the text and language as arguments.

Next, we save the audio file as an MP3 using the save method and give it a name (in this case, audio.mp3). Finally, we play the audio file using the os.system function and the mpg123 command (which is a command-line program for playing MP3 files).

Note that you may need to install the gtts and mpg123 packages before running this code. You can install them using pip install gtts and sudo apt-get install mpg123 (for Linux).